Redundant

+++ Jennings_Tyler_GOL/SeedDialog.cs
@@ -25,7 +25,7 @@ namespace Jennings_Tyler_GOL
             }
             set
             {
-                seedUpDown.Value = (int)value;
+                seedUpDown.Value = value;
             }
         }
         #endregion
